+++ Assets/Scripts/StateMachine/EndGame_ApplicationState.cs
@@ -45,7 +45,7 @@ public class EndGame_ApplicationState : ApplicationState
         Debug.Log("state" + m_type + " ending! ");
     }
 
-    public override int GetMainThemeSchedule()
+    public override float GetMainThemeSchedule()
     {
         return 279;
     }
